subscriber/admonisher/requirements.txt
paho-mqtt==1.6.1
pydub==0.25.1
